+++ R/pkg_scaledic.R
@@ -4,7 +4,6 @@
 #'
 #' @name scaledic-package
 #' @aliases scaledic-package scaledic
-#' @docType package
 #' @author Juergen Wilbert \[aut, cre\]
 #' @keywords package
 #' @importFrom openxlsx write.xlsx
@@ -26,7 +25,7 @@
 #' dat <- impute_missing(dat, scale == "ITRF" & subscale == "Ext")
 #' dat <- impute_missing(dat, scale == "ITRF" & subscale == "Int")
 #' # Show a table with all scales and scale labels included in the data frame
-#' list_scales(dat, labels = TRUE)
+#' list_scales(dat, levels = c("scale_label", "subscale_label"))
 #' # Example with pipeline syntax. Would be much easier to use the "describe" function
 #' # from the psch packages instead of summarise_all here.
 #' library(dplyr)
@@ -36,7 +35,7 @@
 #'   summarise_all(mean, na.rm = TRUE)  |>
 #'   round(2)  |>
 #'   t()
-NULL
+"_PACKAGE"